Add Ancient Game  of Nimm
Nimm is an ancient game of strategy that is named after the old German word for "take." It is also called Tiouk Tiouk in West Africa and Tsynshidzi in China. Players alternate taking stones until there are zero left.

The game of Nimm goes as follows:

   -The game starts with a pile of 20 stones between the players.

   -The two players alternate turns.

   -On a given turn, a player may take either 1 or 2 stone from the center pile.

   -The two players continue until the center pile has run out of stones.

The last player to take a stone loses.